+++ common/src/main/java/com/evandev/remi/mixin/emi/accessor/SizedButtonWidgetAccessor.java
@@ -1,4 +1,4 @@
-package com.evandev.remi.mixin.emi;
+package com.evandev.remi.mixin.emi.accessor;
 
 import dev.emi.emi.screen.widget.SizedButtonWidget;
 import net.minecraft.resources.ResourceLocation;
@@ -7,7 +7,6 @@ import org.spongepowered.asm.mixin.gen.Accessor;
 
 @Mixin(value = SizedButtonWidget.class, remap = false)
 public interface SizedButtonWidgetAccessor {
-
     @Accessor(value = "texture", remap = false)
     void remi$setTexture(ResourceLocation texture);
 }