+++ common/src/main/java/com/evandev/remi/mixin/minecraft/accessor/AbstractContainerScreenAccessor.java
@@ -1,12 +1,16 @@
-package com.evandev.remi.mixin.minecraft;
+package com.evandev.remi.mixin.minecraft.accessor;
 
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
 import net.minecraft.world.inventory.Slot;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;
+import org.spongepowered.asm.mixin.gen.Invoker;
 
 @Mixin(AbstractContainerScreen.class)
 public interface AbstractContainerScreenAccessor {
     @Accessor("hoveredSlot")
     Slot remi$getHoveredSlot();
+
+    @Invoker("findSlot")
+    Slot remi$findSlot(double mouseX, double mouseY);
 }